Rutapp actualizaciones
Ajustes app.py
ajustes de Sweetalert
ajustes vista alertas adminsupadmin

+++ rutapp/app.py
@@ -1,12 +1,23 @@
+# RUTAPP - APLICACIÓN DE GESTIÓN DE RUTAS ESCOLARES
+# Desarrollada por: Cristina Salazar, Camilo Ocampo, Victor Velandia
+
+#==========================================
+# IMPORTACIÓN DE LIBRERÍAS Y CONFIGURACIÓN INICIAL
+#==========================================
+
 from flask import Flask, render_template, request, redirect, url_for,session, flash
 import mysql.connector
 
-app = Flask(__name__)
-app.secret_key = "rutapp_secreto"
+#==========================================
+# CONFIGURACIÓN DE LA APLICACIÓN FLASK
+#==========================================
 
-#============================================
-#Esta es la conexión a la base de datos MySQL
-#============================================
+app = Flask(__name__)
+app.secret_key = "rutapp_secreto"   #   Clave secreta para manejar sesiones y flash messages en Flask
+
+#==========================================
+# CONFIGURACIÓN DE LA CONEXIÓN A LA BASE DE DATOS
+#==========================================
 
 conexion = mysql.connector.connect(
     host="localhost",
@@ -14,6 +25,45 @@ conexion = mysql.connector.connect(
     password="Cristina+-2026",
     database="rutapp_bd"
 )
+#===========================================
+# VARIABLES GLOBALES PARA HTML
+#===========================================
+
+# Aquí se definen variables globales que estarán disponibles en todas las plantillas HTML del proyecto.
+@app.context_processor # Decorador para inyectar variables globales en todas las plantillas HTML
+def inyectar_datos_globales(): # Función que devuelve un diccionario con datos globales disponibles en todas las plantillas
+    nombres_roles = {
+        1: 'Superadministrador',
+        2: 'Administrador',
+        3: 'Conductor',
+        4: 'Padre de familia'           
+    }
+
+    rol = session.get('rol') # Obtiene el rol del usuario desde la sesión
+
+    return {
+        'nombre_panel' : nombres_roles.get(rol, 'Usuario') # Devuelve el nombre del panel según el rol, o 'Usuario' si no se encuentra
+    }
+#Ruta global para regreso al panel
+@app.route('/volver_panel')
+def volver_panel():
+
+    if 'usuario' not in session:
+        return redirect(url_for('login'))
+    
+    rol = session.get('rol')
+
+    if rol == 1:
+        return redirect(url_for('superadministrador'))
+    elif rol == 2:
+        return redirect(url_for('administrador'))
+    elif rol == 3:
+        return redirect(url_for('conductor'))
+    elif rol == 4:
+        return redirect(url_for('padre de familia'))
+    return redirect(url_for('login'))
+
+
 
 #============================================
 #         RUTAS PROYECTO RUTAPP
@@ -63,13 +113,7 @@ def usuarios(): #Esta es la función de Usuarios
 
     cursor.execute(consulta)#Ejecuta la consulta SQL que guardamos en la variable consulta
     lista_usuarios = cursor.fetchall()#Aquí se traen todos los resultados de la consulta.
-
     cursor.close()
-#ruta para boton volver    
-    ruta_volver = {
-        1: 'superadministrador',
-        2: 'administrador'
-    }.get(session['rol'])
 
     return render_template('mod_admin/usuarios.html', usuarios = lista_usuarios)#Manda la lista a usuarios.html
 
@@ -115,8 +159,8 @@ def guardar_usuario():
         cursor.execute(sql, valores)
         conexion.commit()
         cursor.close()
-        flash('Usuario guardado correctamente', 'success')
 
+        flash('Usuario guardado correctamente', 'success')
         return redirect(url_for('usuarios'))
         
 
@@ -176,14 +220,14 @@ def editar_usuario(id_usuario):
     #GET: Cargar formulario
     sql = "SELECT * FROM usuario WHERE id_usuario = %s" # Consulta para obtener los datos actuales del usuario
     cursor.execute(sql, (id_usuario,))
-    usuario = cursor.fetchone()
+    usuario_data = cursor.fetchone()
 
     cursor.close() #Cierra el cursor
 
-    if not usuario: #validacion por si no existe el usuario
+    if not usuario_data: #validacion por si no existe el usuario
         return "Usuario no encontrado", 404
 
-    return render_template('mod_admin/editar_usuario.html', usuario=usuario) # Envía los datos al formulario editar_usuario.html
+    return render_template('mod_admin/editar_usuario.html', usuario=usuario_data) # Envía los datos al formulario editar_usuario.html
   
 
 # ==========================================
@@ -215,13 +259,93 @@ def administrador():
     
     return render_template('mod_admin/admin.html')
 
-#===============ALERTAS DEL SISTEMA=====================#
-@app.route('/gestion_alerta')
-def gestion_alerta():
-    return "<h2>Módulo de alertas en construcción</h2>"
+#============================================================================================
+#===============ALERTAS DEL SISTEMA (mod admin)=====================
+#=============================================================================================
+@app.route('/gestion_alerta')#Rura gestion de alertas Modulo admin y supadmin
+def gestion_alerta():#función para mostrar las alertas en el panel del admin y supadmin
+
+    if 'usuario' not in session:
+        return redirect(url_for('login'))
+
+    if session['rol'] not in [1, 2]:
+        return "Acceso no autorizado"
+
+    cursor = conexion.cursor(dictionary=True)# Crear cursor para interactuar con la base de datos
+
+    # Consulta para traer alertas con datos del conductor, estudiante y ruta
+    consulta = """
+    SELECT 
+        a.*,
+        u.nombres_y_apellidos AS conductor_nombre,
+        e.nombre AS estudiante_nombre,
+        r.nombre_ruta
+    FROM ALERTAS a
+    LEFT JOIN USUARIO u 
+        ON a.id_usuario_emisor = u.id_usuario
+    LEFT JOIN ESTUDIANTE e 
+        ON a.id_estudiante = e.id_estudiante
+    LEFT JOIN RUTA r 
+        ON a.id_ruta = r.id_ruta
+    ORDER BY a.fecha_hora DESC
+    """
+
+    cursor.execute(consulta) #Ejecuta la consulta SQL
+    alertas = cursor.fetchall() #Trae todas las alertas obtenidas de la consulta
+    cursor.close() #Cierra el cursor
+
+    return render_template(
+        'mod_admin/alertas.html', 
+        alertas=alertas
+    ) #Renderiza la plantilla gestion_alerta.html y le pasa la lista de alertas obtenida de la consulta
 
 
-#=========== PANEL CONDUCTOR========================== 
+#===============================================================
+#========Ruta para editar alerta (mod admin y supadmin)=========
+#===============================================================
+
+@app.route('/editar_alerta/<id_alerta>', methods=['GET', 'POST'])
+def editar_alerta(id_alerta):
+    cursor = conexion.cursor(dictionary=True)
+
+    # POST: Guardar cambios en la alerta
+    if request.method == 'POST':
+        estado = request.form.get('estado')
+        observacion = request.form.get('observacion_admin')
+
+        consulta_update = """
+        UPDATE ALERTAS
+        SET estado = %s,
+            observacion_admin = %s
+        WHERE id_alerta = %s
+        """
+
+        cursor.execute(consulta_update, (
+            estado, 
+            observacion, 
+            id_alerta
+        ))
+        conexion.commit()
+        cursor.close()
+
+        flash('Alerta gestionada correctamente', 'success')#nO ESTA FUNCIONANDO REVISAR
+        return redirect(url_for('editar_alerta', id_alerta=id_alerta))
+
+    
+    # GET: Cargar datos actuales de la alerta para mostrar en el formulario
+    consulta = """ SELECT * FROM ALERTAS WHERE id_alerta = %s """
+    cursor.execute(consulta, (id_alerta,))
+    alerta = cursor.fetchone()
+    cursor.close()
+
+    return render_template(# Return de la consulta
+        'mod_admin/editar_alerta.html', 
+        alerta=alerta
+    )
+    
+#============================================================================================
+#============================ PANEL CONDUCTOR==========================
+#============================================================================================
 @app.route('/conductor')
 def conductor():
 
@@ -245,8 +369,9 @@ def mi_ruta():
 def estudiantes_conductor():
     return render_template('mod_conductor/estudiantes_conductor.html')
 
-#===================================================================================================
-# RUTA PARA GESTIONAR ALERTAS DESDE EL PANEL DEL CONDUCTOR
+#==========================================================
+# ======== Ruta para gestionar alertas (conductor)=========
+#=========================================================
 
 @app.route('/conductor/alertas', methods=['GET', 'POST'])#Revisado por Cristina OK#
 def alertas_conductor():
@@ -301,12 +426,13 @@ def alertas_conductor():
     return render_template('mod_conductor/alertas_conductor.html', alertas=alertas)
 
 
-
 @app.route('/conductor/compartir_ubicacion')
 def compartir_ubicacion():  
     return "<h2>Módulo en construcción</h2>"
 
-#============ PANEL PADRES DE FAMILIA==================
+#======================================================================================
+#======================= PANEL PADRES DE FAMILIA==================
+#======================================================================================
 @app.route('/padres')
 def padres():
 
@@ -318,12 +444,89 @@ def padres():
     
     return render_template('mod_padres/padre.html')
 
+#======== Ruta para gestion de alertas (padres)=========
+
+@app.route('/padres/alertas')
+def alertas_padres():
+    cursor = conexion.cursor(dictionary=True)
+
+     #Temporal: para luego conectamos con sesión del padre
+    id_estudiante = 7895462
+
+    consulta = """
+    SELECT *
+    FROM ALERTAS
+    WHERE id_estudiante = %s
+    ORDER BY fecha_hora DESC
+    """
+    cursor.execute(consulta, (id_estudiante,))
+    alertas = cursor.fetchall()
+    cursor.close()
+
+    return render_template(
+        'mod_padres/alertas_padres.html', 
+        alertas=alertas
+    )
+
+#======== Ruta para reporte de inasistencia (padres)=========
+@app.route('/reportar_inasistencia', methods=['GET', 'POST'])
+def reportar_inasistencia():
+    cursor = conexion.cursor(dictionary=True)
+
+    #Temporal: luego conectamos con sesión del padre
+    id_padre = session.get('usuario')
+
+    if request.method == 'POST':
+        estudiante = request.form.get('estudiante')
+        fecha = request.form.get('fecha')
+        motivo = request.form.get('motivo')
+        observacion = request.form.get('observacion')
+
+        mensaje = f"Inasistencia reportada para{estudiante} el día {fecha}. Motivo: {motivo}. Observación{observacion}"
+
+        consulta_insert = """
+        INSERT INTO ALERTAS(
+            id_usuario_emisor,
+            tipo_alerta,
+             mensaje
+        )
+        VALUES (%s, %s, %s)
+        """
+        cursor.execute(consulta_insert, (
+            id_padre,
+            'inasistencia',
+            mensaje
+        ))
+
+        conexion.commit()
+        cursor.close()
+
+        flash('Inasistencia reportada correctamente', 'success')
+        return redirect(url_for('alertas_padres'))
+    cursor.close()
+    return render_template('mod_padres/reportar_inasistencia.html')
+
+
+
 @app.route('/padres/informacion_conductor')
 def informacion_conductor():
     return render_template('mod_padres/informacion_conductor.html')
-#==========================================
-#==========GESTIONAR ESTUDIANTES ==========
-#==========================================
+
+
+@app.route('/estudiantes_padre')
+def estudiantes_padre():
+    return "<h2>Módulo de rutas en construcción</h2>"
+
+
+@app.route('/padres/ver_ruta')
+def ver_ruta():
+    return render_template('mod_padres/ver_ruta.html')
+
+
+
+#===============================================================================
+#==========GESTIONAR ESTUDIANTES (mod_admin mod_supadmin) ==========
+#==================================================================================
 
 @app.route('/gestion_estudiantes')
 def gestion_estudiantes():
@@ -351,14 +554,14 @@ def gestion_estudiantes():
     lista_estudiantes = cursor.fetchall()
     cursor.close()
     
-    ruta_volver = {
-        1: 'superadministrador',
-        2: 'administrador'
-    }.get(session['rol'])
+    # ruta_volver = {
+    #     1: 'superadministrador',
+    #     2: 'administrador'
+    # }.get(session['rol'])
 
     return render_template(
         'mod_admin/estudiantes.html',
-        ruta_volver=ruta_volver,
+        # ruta_volver=ruta_volver,
         estudiantes=lista_estudiantes
     )
 #==== Ruta Crear Estudiante ====
@@ -381,11 +584,14 @@ def guardar_estudiante():
         id_ruta= request.form['id_ruta']
 
         cursor =conexion.cursor()
+
+
+
         # Valida Id dubplicado
         cursor.execute ("SELECT * FROM estudiante WHERE id_estudiante = %s", (id_estudiante,))
         if cursor.fetchone():
             return "Ya existe un estudiante con este número de identificación"
-        
+            
         sql = """
         INSERT INTO estudiante
         (id_estudiante, nombre, grado, direccion, telefono, id_ruta)
@@ -408,11 +614,14 @@ def guardar_estudiante():
 def eliminar_estudiante(id_estudiante):
 
     cursor = conexion.cursor() # Crear cursor
-    sql = "DELETE FROM estudiante WHERE id_estudiante = %s"# Consulta SQL para eliminar el estudiante
-    cursor.execute(sql, (id_estudiante,))# Ejecutar la consulta
-    conexion.commit() # Guardar cambios en la base de datos
-    cursor.close() # Cerrar cursor
-    flash('Estudiante eliminado correctamente', 'success')
+
+    if request.method == 'POST':
+
+        sql = "DELETE FROM estudiante WHERE id_estudiante = %s"# Consulta SQL para eliminar el estudiante
+        cursor.execute(sql, (id_estudiante,))# Ejecutar la consulta
+        conexion.commit() # Guardar cambios en la base de datos
+        cursor.close() # Cerrar cursor
+        flash('Estudiante eliminado correctamente', 'success')
 
     return redirect(url_for('gestion_estudiantes')) # Redirigir a la lista de estudiantes
 
@@ -474,26 +683,7 @@ def editar_estudiante(id_estudiante):
                            ruta_volver=ruta_volver, 
                            estudiante=estudiante  ) # Envía los datos al formulario editar_estudiante.html
 
-#=================================================================
-#=================================================================
 
-@app.route('/reporte_inasistencia')
-def reporte_inasistencia():
-    return "<h2>Módulo de alertas en construcción</h2>"
-
-
-@app.route('/estudiantes_padre')
-def estudiantes_padre():
-    return "<h2>Módulo de rutas en construcción</h2>"
-
-
-@app.route('/padres/ver_ruta')
-def ver_ruta():
-    return render_template('mod_padres/ver_ruta.html')
-
-@app.route('/reportar_inasistencia')
-def reportar_inasistencia():
-    return render_template('mod_padres/reportar_inasistencia.html')
 
 
 #============== RUTA VALIDACION DE LOGIN====================== CRISTINA SALAZAR
